test: simplify cloud signup dependency injection
