+++ test/int/bootstrap/server.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"net/http"
+	"net/netip"
 	"time"
 
 	"github.com/Stackdome/stackdome/cmd/environment"
@@ -48,14 +49,15 @@ func NewServerManager(sessionFactory db.SessionFactory, dbConfig *config.Databas
 	}
 }
 
-func (sm *ServerManager) Bootstrap(ctx context.Context, opts ...environment.EnvConfigOption) error {
+func (sm *ServerManager) Bootstrap(ctx context.Context) error {
 	sm.logger.Info("Starting server bootstrap")
 
-	testEnvironmentOptions := []environment.EnvConfigOption{
+	env := environment.NewTestEnvironment(
+		sm.sessionFactory,
 		environment.WithApplicationConfig(sm.config),
-	}
-	testEnvironmentOptions = append(testEnvironmentOptions, opts...)
-	env := environment.NewTestEnvironment(sm.sessionFactory, testEnvironmentOptions...)
+	)
+	env.Environment().Clients.TurnstileVerifier = noopTurnstileVerifier{}
+	env.Environment().SignupClientIPResolver = fixedSignupClientIPResolver{}
 
 	// Use a background context for environment init. The environment starts
 	// long-lived goroutines (worker manager, cluster manager) that must outlive
@@ -91,6 +93,18 @@ func (sm *ServerManager) Bootstrap(ctx context.Context, opts ...environment.EnvC
 
 	sm.logger.Info("Server bootstrap completed successfully", "baseURL", sm.baseURL)
 	return nil
+}
+
+type noopTurnstileVerifier struct{}
+
+func (noopTurnstileVerifier) Verify(context.Context, string, netip.Addr) error {
+	return nil
+}
+
+type fixedSignupClientIPResolver struct{}
+
+func (fixedSignupClientIPResolver) Resolve(*http.Request) (netip.Addr, error) {
+	return netip.MustParseAddr("127.0.0.1"), nil
 }
 
 func (sm *ServerManager) GetBaseURL() string {
